API Clasificadora/requirements.txt
fastapi
pillow
torch
torchvision